diagnostics: Do not depend on search
