+++ crates/diagnostics/Cargo.toml
@@ -27,7 +27,7 @@ lsp.workspace = true
 markdown.workspace = true
 project.workspace = true
 rand.workspace = true
-search.workspace = true
+
 serde.workspace = true
 serde_json.workspace = true
 settings.workspace = true
