+++ crates/zed_actions/Cargo.toml
@@ -12,4 +12,5 @@ workspace = true
 gpui.workspace = true
 schemars.workspace = true
 serde.workspace = true
+util.workspace = true
 uuid.workspace = true
